+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 ### Fixed
 
 * Performance problem for `hexapdf inspect --pages` when inspecting huge files
-
+* TrueType compound glyph component offset calculation
 
 ## 0.2.0 - 2016-11-28